workingcode/requirements.txt
absl-py==2.1.0
altair==5.0.1
asgiref==3.8.1
astunparse==1.6.3
attrs==23.1.0
audioread==3.0.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.22.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
distlib==0.3.6
Django==5.1.2
djangorestframework==3.15.2
dlib==19.24.6
entry-points-txt==0.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.12.0
flatbuffers==23.5.26
fonttools==4.47.2
fsspec==2023.12.2
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h5py==3.10.0
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
llvmlite==0.42.0
Markdown==3.5.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.7
networkx==3.2.1
numba==0.59.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==23.1
pandas==1.5.2
Pillow==9.4.0
platformdirs==3.5.1
pooch==1.8.0
protobuf==4.23.2
pyarrow==12.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.1.1
pyrsistent==0.19.3
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywin32==305
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.13.2
six==1.16.0
smmap==5.0.0
soundfile==0.12.1
soxr==0.3.7
sqlparse==0.4.3
streamlit==1.23.1
sympy==1.12
tenacity==8.2.2
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch==2.2.0
torchvision==0.17.0
tornado==6.3.2
typing_extensions==4.9.0
tzdata==2022.7
tzlocal==4.3
urllib3==2.0.2
validators==0.20.0
virtualenv==20.23.0
watchdog==3.0.0
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.15.0
